+++ ext/tmdbsimple/configuration.py
@@ -4,8 +4,8 @@
 tmdbsimple.configuration
 ~~~~~~~~~~~~~~~~~~~~~~~~
 
-This module implements the Configuration, Certifications, and Timezones 
-functionality of tmdbsimple.
+This module implements the Configuration and Certifications functionality of
+tmdbsimple.
 
 Created by Celia Oakley on 2013-10-31.
 
@@ -34,7 +34,27 @@ class Configuration(TMDB):
 
     def info(self, **kwargs):
         """
-        Get the system wide configuration info.
+        Get the system wide configuration information. Some elements of the API
+        require some knowledge of this configuration data. The purpose of this
+        is to try and keep the actual API responses as light as possible. It is
+        recommended you cache this data within your application and check for
+        updates every few days.
+
+        This method currently holds the data relevant to building image URLs as
+        well as the change key map.
+
+        To build an image URL, you will need 3 pieces of data. The base_url,
+        size and file_path. Simply combine them all and you will have a fully
+        qualified URL. Here’s an example URL:
+
+            https://image.tmdb.org/t/p/w500/8uO0gUM8aNqYLs1OsTBQiXu0fEv.jpg
+
+        The configuration method also contains the list of change keys which
+        can be useful if you are building an app that consumes data from the
+        change feed.
+
+        Args:
+            None
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -49,6 +69,9 @@ class Configuration(TMDB):
         """
         Get the list of countries (ISO 3166-1 tags) used throughout TMDb.
 
+        Args:
+            None
+
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
@@ -61,6 +84,9 @@ class Configuration(TMDB):
     def jobs(self, **kwargs):
         """
         Get a list of the jobs and departments we use on TMDb.
+
+        Args:
+            None
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -75,6 +101,9 @@ class Configuration(TMDB):
         """
         Get the list of languages (ISO 639-1 tags) used throughout TMDb.
 
+        Args:
+            None
+
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
@@ -88,6 +117,9 @@ class Configuration(TMDB):
         """
         Get a list of the officially supported translations on TMDb.
 
+        Args:
+            None
+
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
@@ -100,6 +132,9 @@ class Configuration(TMDB):
     def timezones(self, **kwargs):
         """
         Get the list of timezones used throughout TMDb.
+
+        Args:
+            None
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -123,23 +158,12 @@ class Certifications(TMDB):
         'tv_list': '/tv/list',
     }
 
-    # here for backward compatability, when only /movie/list existed
-    def list(self, **kwargs):
-        """
-        Get the list of supported certifications for movies.
-
-        Returns:
-            A dict respresentation of the JSON returned from the API.
-        """
-        path = self._get_path('movie_list')
-
-        response = self._GET(path, kwargs)
-        self._set_attrs_to_values(response)
-        return response
-
     def movie_list(self, **kwargs):
         """
         Get an up to date list of the officially supported movie certifications on TMDb.
+
+        Args:
+            None
 
         Returns:
             A dict respresentation of the JSON returned from the API.
@@ -154,6 +178,9 @@ class Certifications(TMDB):
         """
         Get an up to date list of the officially supported TV show certifications on TMDb.
 
+        Args:
+            None
+
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
@@ -163,29 +190,19 @@ class Certifications(TMDB):
         self._set_attrs_to_values(response)
         return response
 
-
-class Timezones(TMDB):
-    """
-    Timezones functionality.
-
-    See: https://developers.themoviedb.org/3/timezones
-    """
-    BASE_PATH = 'timezones'
-    URLS = {
-        'list': '/list',
-    }
-
+    # backward compatability, when only /movie/list existed
     def list(self, **kwargs):
         """
-        Get the list of supported timezones for the API methods that support
-        them.
+        Get the list of supported certifications for movies.
+
+        Args:
+            None
 
         Returns:
             A dict respresentation of the JSON returned from the API.
         """
-        path = self._get_path('list')
+        path = self._get_path('movie_list')
 
         response = self._GET(path, kwargs)
         self._set_attrs_to_values(response)
         return response
-