+++ medusa/server/web/config/general.py
@@ -82,7 +82,6 @@ class ConfigGeneral(Config):
         results = []
 
         # Misc
-        app.DOWNLOAD_URL = download_url
         app.INDEXER_DEFAULT_LANGUAGE = indexerDefaultLang
         app.EP_DEFAULT_DELETED_STATUS = int(ep_default_deleted_status)
         app.SKIP_REMOVED_FILES = config.checkbox_to_value(skip_removed_files)