+++ themes-default/slim/src/components/snatch-selection.vue
@@ -1,5 +1,5 @@
 <template>
-    <div id="snatch-selection-template" class="snatch-selection-template">
+    <div id="snatch-selection-template">
         <vue-snotify />
         <backstretch v-if="show.id.slug" :slug="show.id.slug" />
 
@@ -11,9 +11,9 @@
                      @update-overview-status="filterByOverviewStatus = $event"
         />
 
-        <show-history v-show="show" class="show-history" v-bind="{ show, season, episode }" :key="`history-${show.id.slug}-${season}-${episode || ''}`" />
+        <show-history v-show="show" class="show-history vgt-table-styling" v-bind="{ show, season, episode }" :key="`history-${show.id.slug}-${season}-${episode || ''}`" />
 
-        <show-results v-show="show" class="table-layout" v-bind="{ show, season, episode }" :key="`results-${show.id.slug}-${season}-${episode || ''}`" />
+        <show-results v-show="show" class="table-layout vgt-table-styling" v-bind="{ show, season, episode }" :key="`results-${show.id.slug}-${season}-${episode || ''}`" />
 
     </div>
 </template>
@@ -133,140 +133,27 @@ export default {
 };
 </script>
 
-<style scoped>
-/** Use this as table styling for all table layouts */
-.snatch-selection-template >>> .vgt-table {
-    width: 100%;
-    margin-right: auto;
-    margin-left: auto;
-    text-align: left;
-    border-spacing: 0;
-}
-
-.snatch-selection-template >>> .vgt-table th,
-.snatch-selection-template >>> .vgt-table td {
-    padding: 4px;
-    vertical-align: middle;
-}
-
-/* remove extra border from left edge */
-.snatch-selection-template >>> .vgt-table th:first-child,
-.snatch-selection-template >>> .vgt-table td:first-child {
-    border-left: none;
-}
-
-.snatch-selection-template >>> .vgt-table th {
-    text-align: center;
-    border-collapse: collapse;
-    font-weight: normal;
-}
-
-.snatch-selection-template >>> .vgt-table span.break-word {
-    word-wrap: break-word;
-}
-
-.snatch-selection-template >>> .vgt-table thead th.sorting.sorting-asc {
-    background-position-x: right;
-    background-position-y: bottom;
-}
-
-.snatch-selection-template >>> .vgt-table thead th.sorting {
-    background-repeat: no-repeat;
-}
-
-.snatch-selection-template >>> .vgt-table thead th {
-    padding: 4px;
-    cursor: default;
-}
-
-.snatch-selection-template >>> .vgt-table input.tablesorter-filter {
-    width: 98%;
-    height: auto;
-    -webkit-box-sizing: border-box;
-    -moz-box-sizing: border-box;
-    box-sizing: border-box;
-}
-
-.snatch-selection-template >>> .vgt-table tr.tablesorter-filter-row,
-.snatch-selection-template >>> .vgt-table tr.tablesorter-filter-row td {
-    text-align: center;
-}
-
-/* optional disabled input styling */
-.snatch-selection-template >>> .vgt-table input.tablesorter-filter-row .disabled {
-    display: none;
-}
-
-.tablesorter-header-inner {
-    padding: 0 2px;
-    text-align: center;
-}
-
-.snatch-selection-template >>> .vgt-table tfoot tr {
-    text-align: center;
-    border-collapse: collapse;
-}
-
-.snatch-selection-template >>> .vgt-table tfoot a {
-    text-decoration: none;
-}
-
-.snatch-selection-template >>> .vgt-table th.vgt-row-header {
-    text-align: left;
-}
-
-.snatch-selection-template >>> .vgt-table .season-header {
-    display: inline;
-    margin-left: 5px;
-}
-
-.snatch-selection-template >>> .vgt-table tr.spacer {
-    height: 25px;
-}
-
-.snatch-selection-template >>> .vgt-dropdown-menu {
-    position: absolute;
-    z-index: 1000;
-    float: left;
-    min-width: 160px;
-    padding: 5px 0;
-    margin: 2px 0 0;
-    font-size: 14px;
-    text-align: left;
-    list-style: none;
-    background-clip: padding-box;
-    border-radius: 4px;
-}
-
-.snatch-selection-template >>> .vgt-dropdown-menu > li > span {
-    display: block;
-    padding: 3px 20px;
-    clear: both;
-    font-weight: 400;
-    line-height: 1.42857143;
-    white-space: nowrap;
-}
-
-.snatch-selection-template >>> .align-center {
+<style scoped src="../style/vgt-table.css">
+.vgt-table-styling >>> .align-center {
     display: flex;
     justify-content: center;
 }
 
-.snatch-selection-template >>> .indexer-image :not(:last-child) {
+.vgt-table-styling >>> .indexer-image :not(:last-child) {
     margin-right: 5px;
 }
 
-.snatch-selection-template >>> .button-row {
+.vgt-table-styling >>> .button-row {
     width: 100%;
     display: inline-block;
 }
 
-.snatch-selection-template >>> .fanartBackground {
+.vgt-table-styling >>> .fanartBackground {
     clear: both;
     opacity: 0.9;
 }
 
-.snatch-selection-template >>> .fanartBackground table {
+.vgt-table-styling >>> .fanartBackground table {
     table-layout: auto;
     width: 100%;
     border-collapse: collapse;
@@ -277,7 +164,7 @@ export default {
     color: rgb(0, 0, 0) !important;
 }
 
-.snatch-selection-template >>> .fanartBackground > table th.vgt-row-header {
+.vgt-table-styling >>> .fanartBackground > table th.vgt-row-header {
     border: none !important;
     background-color: transparent !important;
     color: rgb(255, 255, 255) !important;
@@ -285,49 +172,49 @@ export default {
     text-align: left !important;
 }
 
-.snatch-selection-template >>> .fanartBackground td.col-search {
+.vgt-table-styling >>> .fanartBackground td.col-search {
     text-align: center;
 }
 
-.snatch-selection-template >>> .skipped {
+.vgt-table-styling >>> .skipped {
     background-color: rgb(190, 222, 237);
 }
 
-.snatch-selection-template >>> .snatched {
+.vgt-table-styling >>> .snatched {
     background-color: rgb(235, 193, 234);
 }
 
-.snatch-selection-template >>> .downloaded {
+.vgt-table-styling >>> .downloaded {
     background-color: rgb(255, 218, 138);
 }
 
-.snatch-selection-template >>> .failed {
+.vgt-table-styling >>> .failed {
     background-color: rgb(255, 153, 153);
 }
 
-.snatch-selection-template >>> .subtitled {
+.vgt-table-styling >>> .subtitled {
     background-color: rgb(190, 222, 237);
 }
 
-.snatch-selection-template >>> .global-ignored td.release span {
+.vgt-table-styling >>> .global-ignored td.release span {
     color: red;
 }
 
-.snatch-selection-template >>> .show-ignored td.release span {
+.vgt-table-styling >>> .show-ignored td.release span {
     color: red;
     font-style: italic;
 }
 
-.snatch-selection-template >>> .global-required td.release span {
+.vgt-table-styling >>> .global-required td.release span {
     color: green;
 }
 
-.snatch-selection-template >>> .show-required td.release span {
+.vgt-table-styling >>> .show-required td.release span {
     color: green;
     font-style: italic;
 }
 
-.snatch-selection-template >>> .global-undesired td.release span {
+.vgt-table-styling >>> .global-undesired td.release span {
     color: orange;
 }