+++ themes-default/slim/src/store/modules/config/general.js
@@ -7,7 +7,6 @@ const state = {
     donationsUrl: null,
     namingForceFolders: null,
     sourceUrl: null,
-    downloadUrl: null,
     rootDirs: [],
     subtitles: {
         enabled: null
@@ -79,7 +78,10 @@ const state = {
     calendarUnprotected: null,
     calendarIcons: null,
     proxySetting: null,
+    proxyProviders: null,
+    proxyClients: null,
     proxyIndexers: null,
+    proxyOthers: null,
     skipRemovedFiles: null,
     epDefaultDeletedStatus: null,
     developer: null,