+++ themes-default/slim/src/store/modules/config/system.js
@@ -33,6 +33,7 @@ const state = {
     memoryUsage: null,
     schedulers: [],
     showQueue: [],
+    diskSpace: [],
     sslVersion: null,
     pythonVersion: null,
     pid: null,